+++ tests/test_octodns_source_tinydns.py
@@ -17,7 +17,7 @@ class TestTinyDnsFileSource(TestCase):
     def test_populate_normal(self):
         got = Zone('example.com.', [])
         self.source.populate(got)
-        self.assertEqual(17, len(got.records))
+        self.assertEqual(30, len(got.records))
 
         expected = Zone('example.com.', [])
         for name, data in (
@@ -26,8 +26,13 @@ class TestTinyDnsFileSource(TestCase):
                 '',
                 {
                     'type': 'NS',
-                    'ttl': 3600,
-                    'values': ['ns1.ns.com.', 'ns2.ns.com.'],
+                    'ttl': 31,
+                    'values': [
+                        'a.ns.example.com.',
+                        'b.ns.example.com.',
+                        'ns1.ns.com.',
+                        'ns2.ns.com.',
+                    ],
                 },
             ),
             (
@@ -42,6 +47,10 @@ class TestTinyDnsFileSource(TestCase):
             (
                 'cname',
                 {'type': 'CNAME', 'ttl': 3600, 'value': 'www.example.com.'},
+            ),
+            (
+                'cname2',
+                {'type': 'CNAME', 'ttl': 48, 'value': 'www2.example.com.'},
             ),
             (
                 'some-host-abc123',
@@ -61,7 +70,7 @@ class TestTinyDnsFileSource(TestCase):
                             'exchange': 'smtp-1-host.example.com.',
                         },
                         {
-                            'preference': 20,
+                            'preference': 0,
                             'exchange': 'smtp-2-host.example.com.',
                         },
                     ],
@@ -75,11 +84,11 @@ class TestTinyDnsFileSource(TestCase):
                     'values': [
                         {
                             'preference': 30,
-                            'exchange': 'smtp-1-host.example.com.',
+                            'exchange': 'smtp-3-host.mx.example.com.',
                         },
                         {
                             'preference': 40,
-                            'exchange': 'smtp-2-host.example.com.',
+                            'exchange': 'smtp-4-host.mx.example.com.',
                         },
                     ],
                 },
@@ -111,6 +120,83 @@ class TestTinyDnsFileSource(TestCase):
                     'value': 'v=DKIM1\\; k=rsa\\; p=blah',
                 },
             ),
+            ('b.ns', {'type': 'A', 'ttl': 31, 'value': '43.44.45.46'}),
+            ('a.ns', {'type': 'A', 'ttl': 3600, 'value': '42.43.44.45'}),
+            (
+                'smtp-3-host.mx',
+                {'type': 'A', 'ttl': 1800, 'value': '21.22.23.24'},
+            ),
+            (
+                'smtp-4-host.mx',
+                {'type': 'A', 'ttl': 1800, 'value': '22.23.24.25'},
+            ),
+            ('ns5.ns', {'type': 'A', 'ttl': 30, 'value': '14.15.16.17'}),
+            ('ns6.ns', {'type': 'A', 'ttl': 30, 'value': '15.16.17.18'}),
+            (
+                'other',
+                {
+                    'type': 'NS',
+                    'ttl': 30,
+                    'values': ['ns5.ns.example.com.', 'ns6.ns.example.com.'],
+                },
+            ),
+            (
+                '_a._tcp',
+                {
+                    'type': 'SRV',
+                    'ttl': 43,
+                    'values': [
+                        {
+                            'priority': 0,
+                            'weight': 0,
+                            'port': 8888,
+                            'target': 'target.srv.example.com.',
+                        },
+                        {
+                            'priority': 10,
+                            'weight': 50,
+                            'port': 8080,
+                            'target': 'target.somewhere.else.',
+                        },
+                    ],
+                },
+            ),
+            ('target.srv', {'type': 'A', 'ttl': 43, 'value': '56.57.58.59'}),
+            (
+                '_b._tcp',
+                {
+                    'type': 'SRV',
+                    'ttl': 3600,
+                    'values': [
+                        {
+                            'priority': 0,
+                            'weight': 0,
+                            'port': 9999,
+                            'target': 'target.srv.example.com.',
+                        }
+                    ],
+                },
+            ),
+            (
+                'arbitrary-sshfp',
+                {
+                    'type': 'SSHFP',
+                    'ttl': 45,
+                    'values': [
+                        {
+                            'algorithm': 1,
+                            'fingerprint_type': 2,
+                            'fingerprint': '00479b27',
+                        },
+                        {
+                            'algorithm': 2,
+                            'fingerprint_type': 2,
+                            'fingerprint': '00479a28',
+                        },
+                    ],
+                },
+            ),
+            ('arbitrary-a', {'type': 'A', 'ttl': 3600, 'value': '80.81.82.83'}),
         ):
             record = Record.new(expected, name, data)
             expected.add_record(record)
@@ -162,7 +248,10 @@ class TestTinyDnsFileSource(TestCase):
                 {
                     'type': 'PTR',
                     'ttl': 3600,
-                    'value': 'has-dup-def123.example.com.',
+                    'values': [
+                        'has-dup-def123.example.com.',
+                        'has-dup-def456.example.com.',
+                    ],
                 },
             ),
             (
@@ -183,4 +272,5 @@ class TestTinyDnsFileSource(TestCase):
     def test_ignores_subs(self):
         got = Zone('example.com.', ['sub'])
         self.source.populate(got)
-        self.assertEqual(16, len(got.records))
+        # we don't see one www.sub.example.com. record b/c it's in a sub
+        self.assertEqual(29, len(got.records))